Fixing custom provider
